live_transcription/requirements.txt
fastapi
uvicorn
python-dotenv
torch
torchaudio
pyannote.audio
soundfile
numpy
